+++ grocker/__main__.py
@@ -53,10 +53,6 @@ def arg_parser():
         '-r', '--runtime', default=None, choices=('python2', 'python3'),
         help="runtime used to build and run this image.",
     )
-    parser.add_argument(  # precedence
-        '-e', '--entry-point', metavar='<package>', dest='entrypoint', default=None,
-        help="entrypoint used to run this image.",
-    )
     parser.add_argument(
         '--pip-conf', metavar='<file>', type=file_path_or_none_type, default=None,
         help="pip configuration file used to download dependencies (by default use pip config getter).",
@@ -149,7 +145,6 @@ def main(cli_args=None):
     config = parse_config(
         args.config,
         runtime=args.runtime,
-        entrypoint=args.entrypoint,
         pip_constraint=args.pip_constraint,
         docker_image_prefix=args.docker_image_prefix,
     )