+++ tests/resources/grocker-test-project/gtp/__init__.py
@@ -2,4 +2,4 @@
 # Copyright (c) Polyconseil SAS. All rights reserved.
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-__version__ = '1.0.0'
+__version__ = '1.0.1'